Fix circular dependencies
Closes #150
