+++ packages/web3-react/src/components/AlephiumConnect.tsx
@@ -15,7 +15,7 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
-import React, { createContext, useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 
 import defaultTheme from '../styles/defaultTheme'
 
@@ -23,47 +23,12 @@ import AlephiumConnectModal from '../components/ConnectModal'
 import { ThemeProvider } from 'styled-components'
 import { Account, KeyType, SignerProvider } from '@alephium/web3'
 import { Theme, Mode, CustomTheme } from '../types'
-
-export const routes = {
-  CONNECTORS: 'CONNECTORS',
-  PROFILE: 'PROFILE',
-  CONNECT: 'CONNECT'
-}
-
-type Connector = any
-type Error = string | React.ReactNode | null
-
-type ContextValue = {
-  open: boolean
-  setOpen: React.Dispatch<React.SetStateAction<boolean>>
-  route: string
-  setRoute: React.Dispatch<React.SetStateAction<string>>
-  errorMessage: Error
-  connector: Connector
-  setConnector: React.Dispatch<React.SetStateAction<Connector>>
-  account?: Account
-  setAccount: React.Dispatch<React.SetStateAction<Account | undefined>>
-  displayAccount?: (account: Account) => string
-  signerProvider?: SignerProvider
-  setSignerProvider: React.Dispatch<React.SetStateAction<SignerProvider | undefined>>
-  addressGroup?: number
-  keyType?: KeyType
-  network?: string
-  theme: Theme
-  setTheme: React.Dispatch<React.SetStateAction<Theme>>
-  mode: Mode
-  setMode: React.Dispatch<React.SetStateAction<Mode>>
-  customTheme: CustomTheme
-  setCustomTheme: React.Dispatch<React.SetStateAction<CustomTheme>>
-}
-
-const Context = createContext<ContextValue | null>(null)
-
-export const useContext = () => {
-  const context = React.useContext(Context)
-  if (!context) throw Error('AlephiumConnect Hook must be inside a Provider.')
-  return context
-}
+import { routes } from './Common/Modal'
+import {
+  AlephiumConnectContext,
+  AlephiumConnectContextValue,
+  useAlephiumConnectContext
+} from '../contexts/alephiumConnect'
 
 type AlephiumConnectProviderProps = {
   useTheme?: Theme
@@ -86,7 +51,8 @@ export const AlephiumConnectProvider: React.FC<AlephiumConnectProviderProps> = (
 }) => {
   // Only allow for mounting AlephiumConnectProvider once, so we avoid weird global
   // state collisions.
-  if (React.useContext(Context)) {
+  const context = useAlephiumConnectContext()
+  if (context) {
     throw new Error('Multiple, nested usages of AlephiumConnectProvider detected. Please use only one.')
   }
 
@@ -98,7 +64,7 @@ export const AlephiumConnectProvider: React.FC<AlephiumConnectProviderProps> = (
   const [connector, setConnector] = useState<string>('')
   const [route, setRoute] = useState<string>(routes.CONNECTORS)
   const [account, setAccount] = useState<Account>()
-  const [errorMessage, setErrorMessage] = useState<Error>('')
+  const [errorMessage, setErrorMessage] = useState<AlephiumConnectContextValue['errorMessage']>('')
   const [signerProvider, setSignerProvider] = useState<SignerProvider | undefined>()
 
   useEffect(() => setTheme(theme), [theme])
@@ -130,11 +96,11 @@ export const AlephiumConnectProvider: React.FC<AlephiumConnectProviderProps> = (
   }
 
   return (
-    <Context.Provider value={value}>
+    <AlephiumConnectContext.Provider value={value}>
       <ThemeProvider theme={defaultTheme}>
         {children}
         <AlephiumConnectModal theme={theme} mode={mode} customTheme={customTheme} />
       </ThemeProvider>
-    </Context.Provider>
+    </AlephiumConnectContext.Provider>
   )
 }
